+++ Lib/fontbakery/profiles/fontwerk.py
@@ -18,7 +18,6 @@ PROFILE = {
         "fontdata_namecheck",
     ],
     "pending_review": [
-        "glyf_nested_components",
         "vtt_volt_data",  # very similar to vttclean, may be a good idea to merge them.
     ],
     "sections": {